CBD-6342: Black Duck scans for Prometheus cbdeps package
Add configuration to produce Black Duck scans for the product
`cbdep::prometheus`. Ensure that this scan includes the appropriate
version of "Go programming language" that was used to build the cbdeps
package.

Also, modify the Prometheus cbdeps package build process: change the
packaged black-duck-manifest.yaml to specify only `cbdeps::prometheus`
as a component - this will set it up as a Black Duck subproject of
Server.

For both of the above, utilize `NODEJS_VERSION` from the manifest to
install NodeJS.

Also, modify the general Black Duck scan process to understand that
build numbers > 50000 are toy builds. This is necessary so we can scan a
toy build of a new cbdeps package, in order to determine the Black Duck
project ID.

Change-Id: I49734315f3c36fc9dde0788714d86cc469b0071e
Reviewed-on: https://review.couchbase.org/c/build-tools/+/228000
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>